+++ blockfrost-api/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version [next](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.11.0.0...master) (2024-MM-DD)
 
+* Additions [#63](https://github.com/blockfrost/blockfrost-haskell/pull/63)
+  * `/txs/:hash/cbor` endpoint with `TransactionCBOR` data type
+
 # Version [0.11.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.10.0.0...api-0.11.0.0) (2024-08-26)
 
 Changes